HAS-1: Add pagination to search results
This change adds pagination to the search results. After changing the phrase
backend to include pagination info (see [1]), less custom code for the search
results page is required, while automatically benefitting from the existing
code for the pagination. The behaviour is now more similar to the other
collection pages like the archive and the category pages.

[1] https://github.com/wepublish/wepublish/pull/1312