HAS-16: Do not cut articles when visiting from a cafe
This change allows cafes to provide their guests with a secret key that allows
them to read the entire article instead of cutting them off. The key is passed
as a query param at the first visit and stored in a cookie, valid for 4 hours.
When the cookie is set, the entire article is shown. The cookie is checked
whenever a link to an article link is rendered.

The key can be provided by appending ?key=example to the URL of the home page.