HAS-16: Disable paywall even after navigation
In commit 3767a3b4039ab4790c518474fe4946ca7ef55f01, a method to temporarily
disable the paywall was added. When appending a secret key to the URL, the
paywall was to be bypassed for 4 hours.

The change correctly removed the 'articleId' parameter from the URL when
visiting an article, but failed to take into account, that there was a second
condition for hiding the paywall, namely that no navigation must have occurred
before visiting the article.

This change makes sure that the paywall is hidden, even when opening the home
page and then navigating to the article.