Add additional appID to apple-app-site-association for relocated iOS app (#28)
In order for a web page to open in an app instead of the browser, the
associated domain ownership must be proven via the apple-app-site-association
file. After moving the app from Seccom to WePublish, the team id of the app has
changed, requiring a new entry in the association file. This change adds the
new identifier in a backwards compatible way, allowing both versions of the app
to exist at the same time.